+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Overall #
 + Slash commands have been added to the bot.
 + Buttons in the bot are now disabled instead of being deleted.
++ In any reddit command, if there is a video, the thumbnail will be set as the image, and a video link will be provided. This is better than just a spinning gif.
 
 # Events #
 + Markdown syntax for embedding links in text is now detected and, a webhook is sent after deleting the original message.
@@ -30,6 +31,8 @@
 
 # Internet #
 + The youtubenotification, hourlyweather commands have been transferred to slash commands since it is easier with slash commands.
++ Addition of a trivia command.
++ The quote command has been transferred to the internet module. It also uses buttons now.
 
 # Music #
 + The lyrics command's response's title now is the title of the song instead of the query. However, the query is visible as the footer.